# backend/tests/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kivai Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/video/mp4-to-avi/inspect": {
      "post": {
        "summary": "Inspect Mp4 To Avi",
        "operationId": "inspect_mp4_to_avi_video_mp4_to_avi_inspect_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_inspect_mp4_to_avi_video_mp4_to_avi_inspect_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/video/mp4-to-avi": {
      "post": {
        "summary": "Mp4 To Avi",
        "operationId": "mp4_to_avi_video_mp4_to_avi_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_mp4_to_avi_video_mp4_to_avi_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/video/mov-to-mp4/inspect": {
      "post": {
        "summary": "Inspect Mov To Mp4",
        "operationId": "inspect_mov_to_mp4_video_mov_to_mp4_inspect_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_inspect_mov_to_mp4_video_mov_to_mp4_inspect_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/video/mov-to-mp4": {
      "post": {
        "summary": "Mov To Mp4",
        "operationId": "mov_to_mp4_video_mov_to_mp4_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_mov_to_mp4_video_mov_to_mp4_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_inspect_mov_to_mp4_video_mov_to_mp4_inspect_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_inspect_mov_to_mp4_video_mov_to_mp4_inspect_post"
      },
      "Body_inspect_mp4_to_avi_video_mp4_to_avi_inspect_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_inspect_mp4_to_avi_video_mp4_to_avi_inspect_post"
      },
...
      "Body_mov_to_mp4_video_mov_to_mp4_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "default": "auto"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "default": "original"
          },
          "fps": {
            "type": "string",
            "title": "Fps",
            "default": "original"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_mov_to_mp4_video_mov_to_mp4_post"
      },
      "Body_mp4_to_avi_video_mp4_to_avi_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "default": "auto"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "default": "original"
          },
          "fps": {
            "type": "string",
            "title": "Fps",
            "default": "original"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_mp4_to_avi_video_mp4_to_avi_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}